+++ app/global/astro-rpc.js
@@ -9,6 +9,7 @@ function(
     AstroRpc.names = {
         openCart: 'openCart',
         closeCart: 'closeCart',
+        menuItems: 'menuItems',
         showWelcome: 'showWelcome',
         hideWelcome: 'hideWelcome',
         welcomeHasHeader: 'welcomeHasHeader'
@@ -16,6 +17,7 @@ function(
 
     AstroRpc.openCart = Astro.jsRpcMethod(AstroRpc.names.openCart, []);
     AstroRpc.closeCart = Astro.jsRpcMethod(AstroRpc.names.closeCart, []);
+    AstroRpc.menuItems = Astro.jsRpcMethod(AstroRpc.names.menuItems, []);
     AstroRpc.showWelcome = Astro.jsRpcMethod(AstroRpc.names.showWelcome, []);
     AstroRpc.hideWelcome = Astro.jsRpcMethod(AstroRpc.names.hideWelcome, []);
     AstroRpc.welcomeHasHeader = Astro.jsRpcMethod(AstroRpc.names.welcomeHasHeader, []);