+++ app/scaffold-config/baseConfig.js
@@ -1,6 +1,9 @@
 define([], function() {
 /* eslint-enable max-statements */
 
+    // Line altered by generator. If changing please change generator as well!!
+    var iosUsingTabLayout = false;
+
     var colors = {
         primaryColor: '#007ba7',
         secondaryColor: '#007ba7',
@@ -11,6 +14,7 @@ define([], function() {
 
     return {
         colors: colors,
+        iosUsingTabLayout: iosUsingTabLayout,
         loaderColor: loaderColor
     };
 });