+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## To be released
+- iOS layout is now configurable through the `baseConfig.js` file
 - Welcome modal styling is now configurable through the `welcomeConfig.js` file
 - Welcome modal with toggle-able header displays on startup
 - Header content propagated while navigating